WELD-2189 prepare Jigsaw modularisation Moving classes from org.jboss.weld.environment to servlet subpackage to avoid package spliting across artefacts
