--- environments/servlet/core/src/main/java/org/jboss/weld/environment/ContainerContext.java
@@ -1,55 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2010, Red Hat, Inc., and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.weld.environment;
-
-import javax.servlet.ServletContext;
-
-import org.jboss.weld.environment.servlet.WeldServletLifecycle;
-import org.jboss.weld.manager.api.WeldManager;
-import org.jboss.weld.util.Preconditions;
-
-/**
- * Wrap listener arguments.
- *
- * @author <a href="mailto:ales.justin@jboss.org">Ales Justin</a>
- */
-public class ContainerContext {
-
-    private ServletContext servletContext;
-
-    private WeldManager manager;
-
-    public ContainerContext(ServletContext context, WeldManager manager) {
-        Preconditions.checkArgumentNotNull(context, "servlet context");
-
-        this.servletContext = context;
-        if (manager == null) {
-            manager = (WeldManager) context.getAttribute(WeldServletLifecycle.BEAN_MANAGER_ATTRIBUTE_NAME);
-        }
-        this.manager = manager;
-    }
-
-
-    public ServletContext getServletContext() {
-        return servletContext;
-    }
-
-    public WeldManager getManager() {
-        return manager;
-    }
-}
