+++ environments/servlet/core/src/main/java/org/jboss/weld/environment/jetty/JettyContainer.java
@@ -19,8 +19,8 @@ package org.jboss.weld.environment.jetty;
 
 import javax.servlet.ServletContext;
 
-import org.jboss.weld.environment.Container;
-import org.jboss.weld.environment.ContainerContext;
+import org.jboss.weld.environment.servlet.Container;
+import org.jboss.weld.environment.servlet.ContainerContext;
 import org.jboss.weld.environment.servlet.EnhancedListener;
 import org.jboss.weld.environment.servlet.logging.JettyLogger;
 import org.jboss.weld.resources.spi.ResourceLoader;
